requirements.txt
motor~=3.3.1
python-dotenv~=1.0.0
asyncio~=3.4.3
openai~=0.28.1
aiohttp~=3.8.6
aiofiles~=23.2.1
pyTelegramBotAPI